backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
shap==0.43.0
opencv-python==4.8.1.78
Pillow==10.1.0
reportlab==4.0.7
weasyprint==60.1
requests==2.31.0
PyJWT==2.8.1
